+++ tests/integration_server.py
@@ -3,7 +3,6 @@ import os
 import random
 import re
 import shutil
-import six
 import subprocess
 import sys
 import tempfile